refactor: Organize funcs
- Remove callgraph bash script
- Delete func that calls callgraph bash script
- Move gvs.go funcs to appropriate directories
