+++ internal/common/utils.go
@@ -1,7 +1,12 @@
 package common
 
 import (
+	"bytes"
 	"fmt"
+	"log"
+	"os"
+	"os/exec"
+	"path/filepath"
 	"regexp"
 	"strings"
 )
@@ -68,4 +73,64 @@ func UniqueStrings(input []string) []string {
 		}
 	}
 	return result
+}
+
+func CloneRepo(repoURL, branch, cloneDir string) error {
+	os.Setenv("GIT_TERMINAL_PROMPT", "0")
+
+	checkCmd := exec.Command("git", "ls-remote", "--exit-code", repoURL)
+	var checkStderr bytes.Buffer
+	checkCmd.Stderr = &checkStderr
+
+	if err := checkCmd.Run(); err != nil {
+		return fmt.Errorf("repository is not publicly accessible: %s", checkStderr.String())
+	}
+
+	cmd := exec.Command("git", "clone", "--depth", "1", "--branch", branch, "--single-branch", repoURL, cloneDir)
+	var stderr bytes.Buffer
+	cmd.Stderr = &stderr
+
+	err := cmd.Run()
+	if err != nil {
+		return fmt.Errorf("git clone failed: %v\n%s", err, stderr.String())
+	}
+	return nil
+}
+
+func FindGoModDirs(root string) ([]string, error) {
+	var dirs []string
+	err := filepath.WalkDir(root, func(path string, d os.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		if d.IsDir() && d.Name() == "vendor" {
+			return filepath.SkipDir
+		}
+		if d.Name() == "go.mod" {
+			dirs = append(dirs, filepath.Dir(path))
+		}
+		return nil
+	})
+	return dirs, err
+}
+
+func RunGovulncheck(directory, target string) (string, int, error) {
+	cmd := exec.Command("govulncheck", "-format", "sarif", "-C", directory, target)
+	var out bytes.Buffer
+	var stderr bytes.Buffer
+	cmd.Stdout = &out
+	cmd.Stderr = &stderr
+
+	err := cmd.Run()
+	exitCode := 1
+	if cmd.ProcessState != nil {
+		exitCode = cmd.ProcessState.ExitCode()
+	}
+
+	if err != nil && exitCode != 3 {
+		log.Printf("govulncheck failed: %v\nSTDERR:\n%s", err, stderr.String())
+		return stderr.String(), 1, fmt.Errorf("govulncheck failed: %v", err)
+	}
+
+	return out.String(), exitCode, nil
 }
